+++ 1.20/Common/src/main/java/fuzs/proplacer/client/ProPlacerClient.java
@@ -2,9 +2,11 @@ package fuzs.proplacer.client;
 
 import fuzs.proplacer.client.handler.FastPlacementHandler;
 import fuzs.proplacer.client.handler.KeyMappingHandler;
+import fuzs.proplacer.client.handler.ReachAroundPlacementHandler;
 import fuzs.puzzleslib.api.client.core.v1.ClientModConstructor;
 import fuzs.puzzleslib.api.client.core.v1.context.KeyMappingsContext;
 import fuzs.puzzleslib.api.client.event.v1.ClientTickEvents;
+import fuzs.puzzleslib.api.client.event.v1.InteractionInputEvents;
 import fuzs.puzzleslib.api.client.screen.v2.KeyMappingActivationHelper;
 import fuzs.puzzleslib.api.event.v1.LoadCompleteCallback;
 import fuzs.puzzleslib.api.event.v1.entity.player.PlayerInteractEvents;
@@ -18,6 +20,7 @@ public class ProPlacerClient implements ClientModConstructor {
 
     private static void registerEventHandlers() {
         PlayerInteractEvents.USE_BLOCK.register(FastPlacementHandler.INSTANCE::onUseBlock);
+        InteractionInputEvents.USE.register(ReachAroundPlacementHandler::onUseInteraction);
         ClientTickEvents.START.register(FastPlacementHandler.INSTANCE::onStartClientTick);
         ClientTickEvents.START.register(KeyMappingHandler::onStartClientTick);
         LoadCompleteCallback.EVENT.register(KeyMappingHandler::onLoadComplete);