+++ 1.20/Common/src/main/java/fuzs/proplacer/client/handler/KeyMappingHandler.java
@@ -37,7 +37,7 @@ public class KeyMappingHandler {
         isFastPlacementActive = ProPlacer.CONFIG.get(ClientConfig.class).defaultFastPlacement;
     }
 
-    public static boolean isIsFastPlacementActive() {
-        return isFastPlacementActive;
+    public static boolean isFastPlacementActive() {
+        return isFastPlacementActive && ProPlacer.CONFIG.get(ClientConfig.class).allowFastPlacement;
     }
 }