+++ 1.20/Forge/src/main/java/fuzs/proplacer/data/client/ModLanguageProvider.java
@@ -14,7 +14,7 @@ public class ModLanguageProvider extends AbstractLanguageProvider {
     @Override
     protected void addTranslations(TranslationBuilder builder) {
         builder.add(KeyMappingHandler.KEY_CATEGORY, ProPlacer.MOD_NAME);
-        builder.add(KeyMappingHandler.KEY_TOGGLE_FAST_PLACEMENT, "Toggle Fast Placement");
-        builder.add(KeyMappingHandler.KEY_FAST_PLACEMENT_MESSAGE, "Fast Placement: %s");
+        builder.add(KeyMappingHandler.KEY_TOGGLE_FAST_PLACEMENT, "Toggle Fast Block Placement");
+        builder.add(KeyMappingHandler.KEY_FAST_PLACEMENT_MESSAGE, "Fast Block Placement: %s");
     }
 }