+++ 1.20/CHANGELOG.md
@@ -3,7 +3,13 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog].
 
-## [v8.0.0-1.20.1] - 2023-11-07
-- Ported to Minecraft 1.20.1
+## [v8.0.0-1.20.1] - 2024-06-25
+- Fully rewritten for Minecraft 1.20.1
+### Added
+- Add reach-around block placement: place a block in front of the block you are standing on by clicking in midair
+### Changed
+- Placing slabs and stairs is supported much better
+### Fixed
+- Fix missing item use delay for items with a right-click interaction
 
 [Keep a Changelog]: https://keepachangelog.com/en/1.0.0/